requirements.txt
pyrogram
python-decouple
tgcrypto
pytelegrambotapi
sqlalchemy
psycopg2-binary
SQLAlchemy-serializer